Generic Screen Redraw (#1311)
Combine all of the `screen_redraw` implementations into a single generic version. Provide `scene_begin` and `scene_end` for system-specific behavior to still be implemented.
