+++ ENIGMAsystem/SHELL/Graphics_Systems/General/GSscreen.h
@@ -23,6 +23,14 @@
 #include <string>
 using std::string;
 
+namespace enigma {
+  extern unsigned gui_width;
+  extern unsigned gui_height;
+
+  void scene_begin();
+  void scene_end();
+}
+
 namespace enigma_user {
   int screen_save(string filename);
   inline int action_snapshot(string filename)
