--- ENIGMAsystem/SHELL/Graphics_Systems/OpenGL1/GLprofiler.h
@@ -1,24 +1,0 @@
-/** Copyright (C) 2008-2013 Josh Ventura, Robert B. Colton
-***
-*** This file is a part of the ENIGMA Development Environment.
-***
-*** ENIGMA is free software: you can redistribute it and/or modify it under the
-*** terms of the GNU General Public License as published by the Free Software
-*** Foundation, version 3 of the license or any later version.
-***
-*** This application and its source code is distributed AS-IS, WITHOUT ANY
-*** WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
-*** FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
-*** details.
-***
-*** You should have received a copy of the GNU General Public License along
-*** with this code. If not, see <http://www.gnu.org/licenses/>
-**/
-
-// For querying the graphic's output for debugging purposes, eg. triangles rendered, VRAM usage, and VBO's, etc.
-
-#ifndef ENIGMA_GLPROFILER_H
-#define ENIGMA_GLPROFILER_H
-#include "../General/OpenGLHeaders.h"
-
-#endif // ENIGMA_GLPROFILER_H
