+++ ENIGMAsystem/SHELL/Graphics_Systems/OpenGL3/GL3d3d.cpp
@@ -72,9 +72,10 @@ GLenum stenciloperators[8] = {
 namespace enigma_user
 {
 
-void d3d_depth_clear_value(float value) {
+void d3d_clear_depth(double value) {
   oglmgr->BlendFunc();
-  glClearDepthf(value);
+  glClearDepth(value);
+  glClear(GL_DEPTH_BUFFER_BIT);
 }
 
 void d3d_set_software_vertex_processing(bool software) {
