+++ ENIGMAsystem/SHELL/Graphics_Systems/OpenGLES/GLESscreen.cpp
@@ -38,8 +38,6 @@ using namespace std;
   #define GL_BGR 0x80E0
 #endif
 
-extern int window_get_width(), window_get_height();
-
 namespace enigma {
   extern event_iter *event_draw;
 }
@@ -47,99 +45,6 @@ namespace enigma {
 using namespace enigma;
 
 namespace enigma_user {
-
-static inline void draw_back()
-{
-    //Draw backgrounds
-    for (int back_current=0; back_current<7; back_current++)
-    {
-        if (background_visible[back_current] == 1) {
-          if (enigma_user::background_exists(background_index[back_current])) {
-              // if (background_stretched) draw_background_stretched(back, x, y, w, h);
-              draw_background_tiled(background_index[back_current], background_x[back_current], background_y[back_current]);
-          }
-        }
-        // background_foreground, background_index, background_x, background_y, background_htiled,
-        // background_vtiled, background_hspeed, background_vspeed;
-    }
-}
-
-int screen_redraw()
-{
-    if (!view_enabled)
-    {
-       glViewport(0,0,window_get_width(),window_get_height());
-       glLoadIdentity();
-       glScalef(1,-1,1); //OPENGLES
-#if ENIGMA_WS_IPHONE !=0 || ENIGMA_WS_ANDROID !=0
-		glOrthof(0,room_width-1,0,room_height-1,0,1); //OPENGLES put f back in
-#else
-		glOrtho(-1,room_width,-1,room_height,0,1); //OPENGLES put f back in
-#endif
-
-
-      if (background_showcolor)
-      {
-         int clearcolor=((int)background_color)&0xFFFFFF;
-         glClearColor(COL_GET_R(clearcolor)/255.0,COL_GET_G(clearcolor)/255.0,COL_GET_B(clearcolor)/255.0, 1);
-         glClear(GL_COLOR_BUFFER_BIT);
-      }
-        draw_back();
-
-      for (enigma::instance_event_iterator = event_draw->next; enigma::instance_event_iterator != NULL; enigma::instance_event_iterator = enigma::instance_event_iterator->next)
-        enigma::instance_event_iterator->inst->myevent_draw();
-    }
-    else
-    for (view_current=0; view_current<7; view_current++)
-    if (view_visible[(int)view_current])
-    {
-      int vc=(int)view_current;
-      int vob=(int)view_object[vc];
-
-      if (vob != -1)
-      {
-        object_basic *instanceexists = fetch_instance_by_int(vob);
-
-        if (instanceexists)
-        {
-          object_planar* vobr = (object_planar*)instanceexists;
-
-          int vobx=(int)(vobr->x),voby=(int)(vobr->y);
-          //int bbl=*vobr.x+*vobr.bbox_left,bbr=*vobr.x+*vobr.bbox_right,bbt=*vobr.y+*vobr.bbox_top,bbb=*vobr.y+*vobr.bbox_bottom;
-          //if (bbl<view_xview[vc]+view_hbor[vc]) view_xview[vc]=bbl-view_hbor[vc];
-          if (vobx<view_xview[vc]+view_hborder[vc]) view_xview[vc]=vobx-view_hborder[vc];
-          if (vobx>view_xview[vc]+view_wview[vc]-view_hborder[vc]) view_xview[vc]=vobx+view_hborder[vc]-view_wview[vc];
-          if (voby<view_yview[vc]+view_vborder[vc]) view_yview[vc]=voby-view_vborder[vc];
-          if (voby>view_yview[vc]+view_hview[vc]-view_vborder[vc]) view_yview[vc]=voby+view_vborder[vc]-view_hview[vc];
-          if (view_xview[vc]<0) view_xview[vc]=0;
-          if (view_yview[vc]<0) view_yview[vc]=0;
-          if (view_xview[vc]>room_width-view_wview[vc]) view_xview[vc]=room_width-view_wview[vc];
-          if (view_yview[vc]>room_height-view_hview[vc]) view_yview[vc]=room_height-view_hview[vc];
-        }
-      }
-
-      glViewport((int)view_xport[vc],(int)view_yport[vc],(int)view_wport[vc],(int)view_hport[vc]);
-      glLoadIdentity();
-      glScalef(1,-1,1);
-#if ENIGMA_WS_IPHONE !=0 || ENIGMA_WS_ANDROID !=0
-		glOrthof((int)view_xview[vc],(int)view_wview[vc]+(int)view_xview[vc],(int)view_yview[vc],(int)view_hview[vc]+(int)view_yview[vc],0,1); //OPENGLES put f back in
-#else
-      glOrtho((int)view_xview[vc],(int)view_wview[vc]+(int)view_xview[vc],(int)view_yview[vc],(int)view_hview[vc]+(int)view_yview[vc],0,1); //OPENGLES put f back in
-#endif
-      if (background_showcolor)
-      {
-         int clearcolor=((int)background_color)&0xFFFFFF;
-         glClearColor(COL_GET_R(clearcolor)/255.0,COL_GET_G(clearcolor)/255.0,COL_GET_B(clearcolor)/255.0, 1);
-         glClear(GL_COLOR_BUFFER_BIT);
-      }
-      draw_back();
-
-      for (enigma::instance_event_iterator = event_draw->next; enigma::instance_event_iterator != NULL; enigma::instance_event_iterator = enigma::instance_event_iterator->next)
-        enigma::instance_event_iterator->inst->myevent_draw();
-    }
-
-    return 0;
-}
 
 int screen_save(string filename) //Assumes native integers are little endian
 {
